improve readability of the code, add error handling and .gitignore
